Remove deleted blobs from Repository instance variable
The delete_orphaned_blobs method was raising RuntimeErrors as it was trying to remove files that were deleted in previous iterations of the main_loop.

Altered the Repository orphaned_blobs method to be private and added a filenames_to_delete method that yields the filenames to delete, then uses a delete_if on the blobs instance variable to remove the successfully deleted files from the hash.