Ensure errors are raised for any pipeline failures
The command pipelines which are used in `#fetch_manifests` and
`#fetch_jsons` were previously not checking for any command failures.

If `#fetch_jsons` fails but continues the rest of the routine, it means
that blobs that are actually still in use will remain as candidates for
garbage collection.

This appears to be the cause of images being prematurely garbage
collected and appearing as 'failed to solve' errors when trying to use
the images during builds.

By ensuring we raise an error if one of the pipeline commands fails, it
will ensure the GC routine does not continue.